added confirmation alerts to delete button and form submission

+++ server/public/scripts/client.js
@@ -1,6 +1,6 @@
 const taskApp = angular.module('taskApp', []);
 
-taskApp.controller('TaskController', function($http) {
+taskApp.controller('TaskController', function ($http) {
     console.log('Controller is ready');
     const vm = this;
     vm.tasksList = [];
@@ -10,53 +10,57 @@ taskApp.controller('TaskController', function($http) {
         completed: false
     }
 
-    vm.addTask = function(taskIn){
+    vm.addTask = function (taskIn) {
         console.log(taskIn);
-        
-        $http({
-            method: 'POST',
-            url: '/tasks',
-            data: taskIn
-        }).then(function(response){
-            getTasks();
-        }).catch(function(error){
-            alert('Unable to add task!');
-            console.log(error);
-        });
-           
+        let confirmation = confirm('add Task?');
+        if (confirmation === true) {
+            $http({
+                method: 'POST',
+                url: '/tasks',
+                data: taskIn
+            }).then(function (response) {
+                getTasks();
+            }).catch(function (error) {
+                alert('Unable to add task!');
+                console.log(error);
+            });
+        }
     }
 
-    vm.completeTask = function (taskId){
+    vm.completeTask = function (taskId) {
         $http({
             method: 'PUT',
             url: '/tasks/updateTask/' + taskId
-        }).then(function (response){
+        }).then(function (response) {
             getTasks();
-        }).catch(function(error){
+        }).catch(function (error) {
             alert('Unable to update Task!');
             console.log('error', error);
         })
     }
 
-    vm.deleteTask = function(taskId){
-        $http({
-            method: 'DELETE',
-            url: '/tasks/deleteTask/' + taskId
-        }).then(function(response){
-            getTasks();
-        }).catch(function(error){
-            alert('Unable to delete task!');
-            console.log(error);
-        })
+    vm.deleteTask = function (taskId) {
+        let confirmation = confirm('delete task?');
+        if (confirmation === true) {
+            $http({
+                method: 'DELETE',
+                url: '/tasks/deleteTask/' + taskId
+            }).then(function (response) {
+                getTasks();
+            }).catch(function (error) {
+                alert('Unable to delete task!');
+                console.log(error);
+            })
+        }
     }
 
-    function getTasks(){
+    function getTasks() {
         $http({
             method: 'GET',
             url: '/tasks'
-        }).then(function(response){
+        }).then(function (response) {
             vm.tasksList = response.data;
-        }).catch(function(error){
+        }).catch(function (error) {
             alert('Unable to get tasks!');
         })
     }
